Added final files

+++ frontend/src/Components/WorkoutDetails.js
@@ -1,13 +1,24 @@
 import React from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 import { BsTrash } from 'react-icons/bs'
 
+import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+
 const WorkoutDetails = ({workout}) => {
-  const { dispatch } = useWorkoutsContext()
+  const { dispatch } = useWorkoutsContext() 
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+
+    if(!user){
+      return
+    }
     const response = await fetch('/api/workouts/'+workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization' : `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
@@ -21,7 +32,7 @@ const WorkoutDetails = ({workout}) => {
         <h4>{workout.title}</h4>
         <p><strong>Load (kg): </strong>{workout.load}</p>
         <p><strong>Reps: </strong>{workout.reps}</p>
-        <p>{workout.createdAt}</p>
+        <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
         <span onClick={handleClick}><BsTrash/></span>
     </div>
   )
